src/main/java/com/example/seek/utility/FileParser.java
package com.example.seek.utility;

import org.springframework.stereotype.Component;

import java.io.*;
import java.util.ArrayList;
import java.util.List;

@Component
public class FileParser {

    public List<String> parseTrafficCounterFile(String filePath) throws IOException {
        File file = new File(filePath);
        FileInputStream fis = new FileInputStream(file);
        List<String> lines = new ArrayList<>();

        try (BufferedReader br = new BufferedReader(new InputStreamReader(fis))) {
            String line;
            while ((line = br.readLine()) != null) {
                lines.add(line);
            }
        }
        fis.close();

        return lines;
    }
}
